website/server/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
mysqlclient
flask-cors
flask-session
redis